backend/requirements.txt
flask
openai
reportlab
numpy
python-dotenv
pypdf2


